+++ client/views/lists.js
@@ -2,7 +2,7 @@
 
 export default function(discovery) {
     function render(el, config, data, context) {
-        const { item = 'text', limit, emptyText } = config;
+        const { item = 'text', itemConfig, limit, emptyText } = config;
 
         if (emptyText !== false && emptyText !== '') {
             el.setAttribute('emptyText', emptyText || 'Empty list');
@@ -13,10 +13,10 @@ export default function(discovery) {
         }
 
         if (Array.isArray(data)) {
-            discovery.view.renderList(el, {
+            discovery.view.renderList(el, this.composeConfig({
                 view: 'list-item',
                 content: item
-            }, data, context, 0, discovery.view.listLimit(limit, 25));
+            }, itemConfig), data, context, 0, discovery.view.listLimit(limit, 25));
         }
     }