+++ client/views/menu.js
@@ -2,18 +2,18 @@
 
 export default function(discovery) {
     discovery.view.define('menu', function(el, config, data, context) {
-        const { item, limit, emptyText, onClick } = config;
+        const { item, itemConfig, limit, emptyText, onClick } = config;
 
         if (emptyText !== false && emptyText !== '') {
             el.setAttribute('emptyText', emptyText || 'No items');
         }
 
         if (Array.isArray(data)) {
-            discovery.view.renderList(el, {
+            discovery.view.renderList(el, this.composeConfig({
                 view: 'menu-item',
                 content: item,
                 onClick
-            }, data, context, 0, discovery.view.listLimit(limit, 25));
+            }, itemConfig), data, context, 0, discovery.view.listLimit(limit, 25));
         }
     });
 }